Fix proto_go_library deps resolution for "merged rules"
If two proto_go_library rules share the same importpath, the srcs and deps
for the two rules are added rather than creating a new rule.  This fixes
a bug whereby the resolver was being populated only for the "merge destination"
rule and not for any "merge source" rules.  google/api/field_behaviour.proto
is an example of a "merge source" type rule, and is added to the proto_repository test.